Handle issue and pull request in the same template

+++ lib/Gitprep/API.pm
@@ -524,38 +524,37 @@ sub api_update_issue_message {
   my $issue_message = $self->app->dbi->model('issue_message')->select(
     {user => ['id']}, where => {'issue_message.row_id' => $issue_message_row_id}
   )->one;
-  
+
+  my $json = {success => 0};
   my $session_user_id = $self->session_user_id;
 
-  my $is_my_project = $user_id eq $session_user_id;
-  my $is_my_comment = $issue_message->{'user.id'} eq $session_user_id;
-  my $can_modify = $is_my_project || $is_my_comment;
-  
-  my $json;
-  if ($can_modify) {
-    my $now_tm = Time::Moment->now;
-    my $update_time = $now_tm->epoch;
-    $self->app->log->info($update_time);
+  if ($session_user_id) {
+    my $is_my_project = $user_id eq $session_user_id;
+    my $is_my_comment = $issue_message->{'user.id'} eq $session_user_id;
+    my $can_modify = $is_my_project || $is_my_comment;
+
+    if ($can_modify) {
+      my $now_tm = Time::Moment->now;
+      my $update_time = $now_tm->epoch;
+      $self->app->log->info($update_time);
     
-    $self->app->dbi->model('issue_message')->update(
-      {
-        message => $message,
-        update_time => $update_time
-      },
-      where => {row_id => $issue_message_row_id}
-    );
-    
-    my $markdown_message = $self->markdown($message);
-    
-    $json = {
-      success => 1,
-      markdown_message => $markdown_message
-    };
+      $self->app->dbi->model('issue_message')->update(
+        {
+          message => $message,
+          update_time => $update_time
+        },
+        where => {row_id => $issue_message_row_id}
+      );
+
+      my $markdown_message = $self->markdown($message);
+
+      $json = {
+        success => 1,
+        markdown_message => $markdown_message
+      };
+    }
   }
-  else {
-    $json = {success => 0};
-  }
-  
+
   return $json;
 }
 
@@ -565,25 +564,24 @@ sub api_delete_issue_message {
   my $issue_message = $self->app->dbi->model('issue_message')->select(
     {user => ['id']}, where => {'issue_message.row_id' => $issue_message_row_id}
   )->one;
-  
+
+  my $json = {success => 0};
   my $session_user_id = $self->session_user_id;
 
-  my $is_my_project = $user_id eq $session_user_id;
-  my $is_my_comment = $issue_message->{'user.id'} eq $session_user_id;
-  my $can_modify = $is_my_project || $is_my_comment;
-  
-  my $json;
-  if ($can_modify) {
-    $self->app->dbi->model('issue_message')->delete(
-      where => {row_id => $issue_message_row_id}
-    );
-    
-    $json = {success => 1};
+  if ($session_user_id) {
+    my $is_my_project = $user_id eq $session_user_id;
+    my $is_my_comment = $issue_message->{'user.id'} eq $session_user_id;
+    my $can_modify = $is_my_project || $is_my_comment;
+
+    if ($can_modify) {
+      $self->app->dbi->model('issue_message')->delete(
+        where => {row_id => $issue_message_row_id}
+      );
+
+      $json = {success => 1};
+    }
   }
-  else {
-    $json = {success => 0};
-  }
-  
+
   return $json;
 }
